fix winrepo_source_dir documentation
The winrepo_source_dir minion setting is valid if the minion is in master-minion mode or in masterless mode.  Currently it is documented that the only way you can use winrepo_source_dir is in masterless mode.  I updated the documentation to include a new "Minion Windows Software Repo Settings" section and moved winrepo_source_dir to this new section.